fix: Fix failed test cases
