+++ packages/mip/test/specs/register-element.spec.js
@@ -3,13 +3,13 @@
  * @author huanghuiquan(huanghuiquan@baidu.com)
  */
 
-import registerElement from 'src/register-element.js'
-import CustomElement from 'src/custom-element.js'
-import store from 'src/custom-element-store.js'
+import registerElement from 'src/register-element'
+import CustomElement from 'src/custom-element'
+import store from 'src/custom-element-store'
 import cssLoader from 'src/util/dom/css-loader'
 import performance from 'src/performance'
 
-describe('Register element', function () {
+describe('register-element', () => {
   let prefix = 'mip-test-register-element'
   class MIPExample extends CustomElement {
     static get observedAttributes () {
@@ -17,9 +17,7 @@ describe('Register element', function () {
     }
   }
 
-  let createElement = (tag) => document.createElement(tag)
-
-  it('twice should define customElements once', function () {
+  it('should define custom element once', () => {
     let get = sinon.spy(store, 'get')
     let set = sinon.spy(store, 'set')
     let def = sinon.spy(window.customElements, 'define')
@@ -38,7 +36,7 @@ describe('Register element', function () {
     expect(def.calledOnce).to.be.true
   })
 
-  it('should load css if pass css text', function () {
+  it('should load css if pass css text', () => {
     let insertStyleElement = sinon.spy(cssLoader, 'insertStyleElement')
     let name = prefix + '-css'
     let css = 'body {color: red;}'
@@ -66,7 +64,7 @@ describe('Register element', function () {
 
     registerElement(name, MIPExample)
 
-    let ele = createElement(name)
+    let ele = document.createElement(name)
     let lifecycSpies = lifecycs.map(cbName => sinon.spy(ele.customElement, cbName))
 
     ele.setAttribute('name', 'fake')
@@ -83,7 +81,7 @@ describe('Register element', function () {
     let name = prefix + 'add-class'
     registerElement(name, MIPExample)
 
-    let ele = createElement(name)
+    let ele = document.createElement(name)
 
     document.body.appendChild(ele)
     document.body.removeChild(ele)
@@ -101,7 +99,7 @@ describe('Register element', function () {
       }
     })
 
-    let ele = createElement(name)
+    let ele = document.createElement(name)
     expect(ele.customElement.build).to.throw('build error')
 
     document.body.appendChild(ele)
@@ -126,257 +124,12 @@ describe('Register element', function () {
       }
     })
 
-    let ele = createElement(name)
+    let ele = document.createElement(name)
     document.body.appendChild(ele)
     document.body.removeChild(ele)
 
     addFsElement.restore()
 
     sinon.assert.calledOnce(addFsElement)
-  })
-
-  describe('placeholder', function () {
-    let name = prefix + 'placeholder'
-    let ele
-    registerElement(name, class extends MIPExample {
-      createPlaceholderCallback () {
-        let loader = document.createElement('div')
-        loader.classList.add('default-placeholder')
-        return loader
-      }
-    })
-
-    beforeEach(() => {
-      ele = createElement(name)
-    })
-
-    afterEach(() => {
-      ele.parentElement && ele.parentElement.removeChild(ele)
-    })
-
-    it('should show default placeholder if not placeholder attribute', async function () {
-      ele.setAttribute('layout', 'responsive')
-      ele.setAttribute('width', '1')
-      ele.setAttribute('height', '1')
-
-      document.body.appendChild(ele)
-
-      ele.viewportCallback(true)
-
-      let placeholder = ele.querySelector('.default-placeholder')
-      expect(placeholder).to.not.be.null
-    })
-
-    it('should show placeholder if element has placeholder attribute', async function () {
-      ele.setAttribute('layout', 'responsive')
-      ele.setAttribute('width', '1')
-      ele.setAttribute('height', '1')
-
-      let placeholder = createElement('div')
-      placeholder.setAttribute('placeholder', '')
-      ele.appendChild(placeholder)
-      document.body.appendChild(ele)
-
-      ele.viewportCallback(true)
-      let eleRect = ele.getBoundingClientRect()
-      let placeholderRect = placeholder.getBoundingClientRect()
-      expect(eleRect).to.deep.equal(placeholderRect)
-      expect(placeholder.classList.contains('mip-hidden')).to.be.false
-
-      await new Promise(resolve => (ele.customElement.firstLayoutCompleted = resolve))
-      expect(placeholder.classList.contains('mip-hidden')).to.be.true
-      ele.customElement.togglePlaceholder(true)
-      expect(placeholder.classList.contains('mip-hidden')).to.be.false
-    })
-
-    it('should ignore input placeholder if element has placeholder attribute', async function () {
-      let placeholder = createElement('input')
-      placeholder.setAttribute('placeholder', 'ddd')
-      ele.appendChild(placeholder)
-      document.body.appendChild(ele)
-
-      ele.viewportCallback(true)
-      let eleRect = ele.getBoundingClientRect()
-      let placeholderRect = placeholder.getBoundingClientRect()
-      expect(eleRect).to.not.equal(placeholderRect)
-      expect(placeholder.classList.contains('mip-hidden')).to.be.false
-
-      await new Promise(resolve => (ele.customElement.firstLayoutCompleted = resolve))
-      expect(placeholder.classList.contains('mip-hidden')).to.be.false
-    })
-
-    it('should not show placeholder if self is a placeholder', async function () {
-      let placeholder = createElement('div')
-      placeholder.setAttribute('placeholder', '')
-      ele.setAttribute('placeholder', '')
-      ele.appendChild(placeholder)
-      document.body.appendChild(ele)
-
-      ele.viewportCallback(true)
-      let eleRect = ele.getBoundingClientRect()
-      let placeholderRect = placeholder.getBoundingClientRect()
-      expect(eleRect).to.not.equal(placeholderRect)
-      expect(placeholder.classList.contains('mip-hidden')).to.be.false
-
-      await new Promise(resolve => (ele.customElement.firstLayoutCompleted = resolve))
-      expect(placeholder.classList.contains('mip-hidden')).to.be.false
-    })
-  })
-
-  describe('loading', function () {
-    let name = prefix + 'loading'
-    let ele
-    class MIPLoadingExample extends MIPExample {
-      isLoadingEnabled () {
-        return true
-      }
-    }
-    registerElement(name, MIPLoadingExample)
-
-    beforeEach(() => {
-      ele = createElement(name)
-    })
-
-    afterEach(() => {
-      ele.parentElement && ele.parentElement.removeChild(ele)
-    })
-
-    it('should show loading if extensions enable loading', async function () {
-      ele.setAttribute('layout', 'responsive')
-      ele.setAttribute('width', '1')
-      ele.setAttribute('height', '1')
-      document.body.appendChild(ele)
-
-      expect(ele.querySelector('.mip-loading-container')).to.be.null
-
-      ele.viewportCallback(true)
-
-      // show loading
-      let loader = ele.querySelector('.mip-loading-container')
-      expect(loader).to.be.not.null
-      let eleRect = ele.getBoundingClientRect()
-      let loaderRect = loader.getBoundingClientRect()
-      expect(eleRect).to.not.equal(loaderRect)
-
-      await Promise.resolve()
-      expect(ele.querySelector('.mip-loading-container')).to.be.null
-    })
-
-    it('should not show loading if element implict noloading', async function () {
-      ele.setAttribute('layout', 'responsive')
-      ele.setAttribute('width', '1')
-      ele.setAttribute('height', '1')
-      ele.setAttribute('noloading', '')
-      document.body.appendChild(ele)
-
-      expect(ele.querySelector('.mip-loading-container')).to.be.null
-
-      ele.viewportCallback(true)
-
-      // show loading
-      let loader = ele.querySelector('.mip-loading-container')
-      expect(loader).to.be.null
-    })
-  })
-
-  describe('fallback', function () {
-    let name = prefix + 'fallback'
-    let ele
-    registerElement(name, MIPExample)
-
-    beforeEach(() => {
-      ele = createElement(name)
-    })
-
-    afterEach(() => {
-      ele.parentElement && ele.parentElement.removeChild(ele)
-    })
-
-    it('should show fallback element if declare in html and load resources error', async function () {
-      ele.setAttribute('layout', 'responsive')
-      ele.setAttribute('width', '1')
-      ele.setAttribute('height', '1')
-
-      let fallback = createElement('div')
-      fallback.setAttribute('fallback', '')
-      ele.appendChild(fallback)
-      document.body.appendChild(ele)
-
-      await new Promise(resolve => {
-        // override to make it reject
-        ele.customElement.layoutCallback = function (err) {
-          resolve()
-          return Promise.reject(err)
-        }
-        ele.viewportCallback(true)
-      })
-
-      let eleRect = ele.getBoundingClientRect()
-      let fallbackRect = fallback.getBoundingClientRect()
-      expect(eleRect).to.deep.equal(fallbackRect)
-      expect(fallback.classList.contains('mip-hidden')).to.be.false
-    })
-  })
-
-  describe('applySizesAndMediaQuery', function () {
-    let name = prefix + 'size-and-media-query'
-    let ele
-    let ww = window.innerWidth
-    registerElement(name, MIPExample)
-
-    beforeEach(() => {
-      ele = createElement(name)
-    })
-
-    afterEach(() => {
-      document.body.removeChild(ele)
-    })
-
-    it('normal sizes not match', async function () {
-      ele.setAttribute('sizes', `(min-width: ${ww + 1}px) 50vw, 100vw`)
-      document.body.appendChild(ele)
-      expect(ele.style.width).to.equal('100vw')
-    })
-
-    it('normal sizes match', async function () {
-      ele.setAttribute('sizes', `(min-width: ${ww - 1}px) 50vw, 100vw`)
-      document.body.appendChild(ele)
-      expect(ele.style.width).to.equal('50vw')
-    })
-
-    it('normal heights not match', async function () {
-      ele.setAttribute('sizes', `(min-width: ${ww + 1}px) 50vw, 100vw`)
-      document.body.appendChild(ele)
-      expect(ele.style.width).to.equal('100vw')
-    })
-
-    it('normal heights match', async function () {
-      ele.setAttribute('sizes', `(min-width: ${ww - 1}px) 50vw, 100vw`)
-      document.body.appendChild(ele)
-      expect(ele.style.width).to.equal('50vw')
-    })
-
-    it('normal media query match', async function () {
-      ele.setAttribute('media', `(min-width: ${ww - 1}px)`)
-      document.body.appendChild(ele)
-      expect(ele.classList.contains('mip-hidden-by-media-query')).to.be.false
-    })
-
-    it('normal media query not match', async function () {
-      ele.setAttribute('media', `(min-width: ${ww + 1}px)`)
-      document.body.appendChild(ele)
-      expect(ele.classList.contains('mip-hidden-by-media-query')).to.be.true
-    })
-
-    it('normal heights width layout=responsive', async function () {
-      ele.setAttribute('layout', 'responsive')
-      ele.setAttribute('width', '1px')
-      ele.setAttribute('height', '1px')
-      ele.setAttribute('heights', `(min-width: ${ww + 1}px) 50vw, 100vw`)
-      console.log('ddd')
-      document.body.appendChild(ele)
-
-      expect(ele.querySelector('mip-i-space').style.paddingTop).to.equal('100vw')
-    })
   })
 })
